+++ tcod/cffi.h
@@ -1,5 +1,5 @@
 /* This header is the entry point for the cffi parser.
-   Anything included here will be accessible from tcod.libtcod.lib */
+   Anything included here will be accessible from tcod.loader.lib */
 #include "../libtcod/src/libtcod/libtcod.h"
 #include "../libtcod/src/libtcod/libtcod_int.h"
 #include "../libtcod/src/libtcod/wrappers.h"