+++ tcod/map.py
@@ -7,7 +7,7 @@ import warnings
 
 import numpy as np
 
-from tcod.libtcod import lib, ffi
+from tcod.loader import lib, ffi
 import tcod._internal
 import tcod.constants