+++ tcod/noise.py
@@ -38,7 +38,7 @@ from typing import Any, Optional
 import numpy as np
 
 from tcod._internal import deprecate
-from tcod.libtcod import ffi, lib
+from tcod.loader import ffi, lib
 import tcod.constants
 import tcod.random